Perf: seg layers int32 -> uint16 (halves resident memory)
All three class layers were allocated int32 while instance ids are
already hard-capped at 65535 by the uint16 TIF export. On a
1000-frame 512x512 stack that was ~3.1 GB of layers; uint16 makes it
~1.6 GB. next_instance_id now raises a clear error at the 65535
ceiling instead of letting a larger id silently wrap in the array.
The transient int32 copies handed to the tracker and SAM remain int32
(their consumers expect plain integer label arrays).

Round-trip tested offscreen: paint -> fill undo/redo -> delete undo ->
save (uint16 TIF) -> session reload -> overlay render, plus the
ceiling guard.

+++ controllers/tool_controller.py
@@ -3031,7 +3031,7 @@ class ToolController:
         ct = ct.lower()
 
         try:
-            arr = mask_io.load_mask_tif(path).astype(np.int32)
+            arr = mask_io.load_mask_tif(path).astype(np.uint16)
         except Exception as e:
             QMessageBox.critical(self.window, "Load Class",
                                   f"Failed to read TIF:\n\n{e}")
@@ -3449,7 +3449,7 @@ class ToolController:
                 raise ValueError(
                     f"{ct} layer shape {arr.shape} does not match "
                     f"cell shape {(T, H, W)}")
-            seg.set_layer(ct, arr.astype(np.int32))
+            seg.set_layer(ct, arr.astype(np.uint16))
             for iid in np.unique(arr):
                 if iid == 0:
                     continue
